api_gateway/requirements.txt
redis==5.2.1
fastapi==0.115.6
async_timeout==5.0.1
fastapi_gateway==0.0.4
uvicorn==0.20.0
httpx==0.28.1
pytest==8.3.4
pytest-dependency==0.6.0
pyjwt[crypto]==2.10.1
